+++ src/main/knowledage/loader/noteLoader.ts
