+++ src/main/services/KnowledgeService.ts
@@ -21,10 +21,10 @@ import type { ExtractChunkData } from '@cherrystudio/embedjs-interfaces'
 import { LibSqlDb } from '@cherrystudio/embedjs-libsql'
 import { SitemapLoader } from '@cherrystudio/embedjs-loader-sitemap'
 import { WebLoader } from '@cherrystudio/embedjs-loader-web'
-import Embeddings from '@main/embeddings/Embeddings'
-import { addFileLoader } from '@main/loader'
-import { NoteLoader } from '@main/loader/noteLoader'
-import Reranker from '@main/reranker/Reranker'
+import Embeddings from '@main/knowledage/embeddings/Embeddings'
+import { addFileLoader } from '@main/knowledage/loader'
+import { NoteLoader } from '@main/knowledage/loader/noteLoader'
+import Reranker from '@main/knowledage/reranker/Reranker'
 import { windowService } from '@main/services/WindowService'
 import { getDataPath } from '@main/utils'
 import { getAllFiles } from '@main/utils/file'
